+++ semiteam3/src/main/java/com/kh/semiteam3/dao/ReplyDao.java
@@ -68,7 +68,7 @@ public class ReplyDao {
 	public List<ReplyDto> findBylist(String memberId, PageVO pageVO) {
 	    String sql = "select * from (" 
 	            + "select rownum rn, TMP.* from (" 
-	            + "select reply.reply_no, reply.reply_content, reply.reply_time, member.member_id, "
+	            + "select reply.reply_no, reply.reply_content, reply.reply_time, reply.board_no, member.member_id, "
 	            + "board.board_view, board.board_title, board.board_reply "
 	            + "from reply "
 	            + "inner join member "
